+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+Features:
+
+- Support configuring the instance edition. Defaults to standard, unless autoscaling is enabled, in which case it defaults to enterprise.
+
 ## v0.3.2 (2025-09-12)
 
 Chores: